Remove eucovidcert route, controller, service and tests Remove feature flag and env variables Remove @pagopa/io-functions-eucovidcerts-sdk package
