+++ .env.example
@@ -69,14 +69,6 @@ CGN_OPERATOR_SEARCH_API_URL=http://host.docker.internal:7074
 CGN_OPERATOR_SEARCH_API_BASE_PATH="/api/v1/cgn/operator-search"
 
 # ------------------------------------
-# EuCovidCert Env Variables
-# ------------------------------------
-FF_EUCOVIDCERT_ENABLED=1
-EUCOVIDCERT_API_KEY=put_your_api_key_here
-EUCOVIDCERT_API_URL=http://host.docker.internal:7071/api/v1
-EUCOVIDCERT_API_BASE_PATH="/api/v1/eucovidcert"
-
-# ------------------------------------
 # PEC SERVER support env Variables
 # ------------------------------------
 PECSERVERS_poste_url="http://io-mvl-server-mock:4000"
