+++ helm/values-prod01.yaml
@@ -59,8 +59,6 @@ microservice-chart:
     CGN_OPERATOR_SEARCH_API_BASE_PATH: '/api/v1/cgn/operator-search'
     CGN_OPERATOR_SEARCH_API_URL: 'https://cgnonboardingportal-p-op.azurewebsites.net'
     DIAGNOSTICS_AZUREBLOBRETENTIONINDAYS: '90'
-    EUCOVIDCERT_API_BASE_PATH: '/api/v1/eucovidcert'
-    EUCOVIDCERT_API_URL: 'https://io-p-eucovidcert-fn.azurewebsites.net/api/v1'
     FETCH_KEEPALIVE_ENABLED: 'true'
     FETCH_KEEPALIVE_FREE_SOCKET_TIMEOUT: '30000'
     FETCH_KEEPALIVE_MAX_FREE_SOCKETS: '10'
@@ -69,7 +67,6 @@ microservice-chart:
     FETCH_KEEPALIVE_TIMEOUT: '60000'
     FF_BONUS_ENABLED: '1'
     FF_CGN_ENABLED: '1'
-    FF_EUCOVIDCERT_ENABLED: '1'
     FF_IO_SIGN_ENABLED: '1'
     FF_PN_ACTIVATION_ENABLED: '1'
     FF_ROUTING_PUSH_NOTIF: 'ALL'
@@ -118,7 +115,6 @@ microservice-chart:
     CGN_API_KEY: 'funccgn-KEY-APPBACKEND'
     IO_SIGN_API_KEY: 'funciosign-KEY-APPBACKEND'
     CGN_OPERATOR_SEARCH_API_KEY: 'funccgnoperatorsearch-KEY-PROD-APPBACKEND'
-    EUCOVIDCERT_API_KEY: 'funceucovidcert-KEY-APPBACKEND'
     APP_MESSAGES_API_KEY: 'appbackend-APP-MESSAGES-API-KEY'
     PRE_SHARED_KEY: 'appbackend-PRE-SHARED-KEY'
     PAGOPA_API_KEY_PROD: 'appbackend-PAGOPA-API-KEY-PROD-PRIMARY'
