--- openapi/api_eucovidcert.template.yaml
@@ -1,79 +1,0 @@
-swagger: "2.0"
-info:
-  version: 16.7.0
-  title: EU Covid Certificate service for IO App
-basePath: /api/v1/eucovidcert
-schemes:
-  - https
-security:
-  - Bearer: []
-paths:
-  /certificate:
-    post:
-      operationId: getCertificate
-      summary: Retrieve a certificate for a given Citizen
-      description: |-
-        Given a Citizen's fiscal code and an OTP (the auth code previously sent via IO Message), a Certificate is returned with ID, encoded QR-code, expiration date and a markdown text with all meaningful information to be shown to Citizens.
-      parameters:
-        - name: accessData
-          in: body
-          schema:
-            $ref: "#/definitions/GetCertificateParams"
-      responses:
-        '200':
-          description: |-
-            A Certificate exists and it's found for the given access data. It is retrieved regardless of it's expired or its current status
-          schema:
-            $ref: "#/definitions/Certificate"
-        '400':
-          description: Payload has bad format
-          schema:
-            $ref: "#/definitions/ProblemJson"
-        '401':
-          description: Bearer token null or expired
-        '403':
-          description: Access data provided are invalid or no Certificate has been emitted for the given Citizen
-        '410':
-          description: Endpoint no longer available
-        '500':
-          description: Generic server error
-          schema:
-            $ref: "#/definitions/ProblemJson"
-        '504':
-          description: Gateway Timeout
-definitions:
-  GetCertificateParams:
-    type: object
-    properties:
-      auth_code:
-        type: string
-        description: OTP previously sent to the Citizen
-      preferred_languages:
-        $ref: '#/definitions/PreferredLanguages'
-    required:
-      - auth_code
-  Certificate:
-    $ref: "../node_modules/@pagopa/io-functions-eucovidcerts-sdk/openapi.yaml#/definitions/Certificate"
-  ValidCertificate:
-    $ref: "../node_modules/@pagopa/io-functions-eucovidcerts-sdk/openapi.yaml#/definitions/ValidCertificate"
-  RevokedCertificate:
-    $ref: "../node_modules/@pagopa/io-functions-eucovidcerts-sdk/openapi.yaml#/definitions/RevokedCertificate"
-  ExpiredCertificate:
-    $ref: "../node_modules/@pagopa/io-functions-eucovidcerts-sdk/openapi.yaml#/definitions/ExpiredCertificate"
-  HeaderInfo:
-    $ref: "../node_modules/@pagopa/io-functions-eucovidcerts-sdk/openapi.yaml#/definitions/HeaderInfo"
-  QRCode:
-    $ref: "../node_modules/@pagopa/io-functions-eucovidcerts-sdk/openapi.yaml#/definitions/QRCode"
-  PreferredLanguages:
-    $ref: "../node_modules/@pagopa/io-functions-eucovidcerts-sdk/openapi.yaml#/definitions/PreferredLanguages"
-  ProblemJson:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/ProblemJson"
-consumes:
-  - application/json
-produces:
-  - application/json
-securityDefinitions:
-  Bearer:
-    type: apiKey
-    name: Authorization
-    in: header
