+++ package.json
@@ -52,7 +52,6 @@
     "generate:api:pagopaproxy": "rimraf generated/pagopa-proxy && gen-api-models --api-spec https://raw.githubusercontent.com/pagopa/io-pagopa-proxy/v1.6.0/api-spec/api-for-io.yaml --no-strict --out-dir generated/pagopa-proxy --request-types --response-decoders --client",
     "generate:api:lollipop": "rimraf generated/lollipop-api && gen-api-models --api-spec https://raw.githubusercontent.com/pagopa/io-functions-lollipop/724c3ba79d272d4e2555dbc2a8e1657bec9f84ce/openapi/internal.yaml --no-strict --out-dir generated/lollipop-api --request-types --response-decoders --client",
     "generate:api:services-app-backend": "rimraf generated/services-app-backend && gen-api-models --api-spec https://raw.githubusercontent.com/pagopa/io-services-cms/30efc51d50e72685d2cf7739967ca4a6b36037f5/apps/app-backend/api/internal.yaml --no-strict --out-dir generated/services-app-backend --request-types --response-decoders --client",
-    "generate:proxy:eucovidcert-models": "rimraf generated/eucovidcert && gen-api-models --api-spec api_eucovidcert.yaml --out-dir generated/eucovidcert",
     "generate:proxy:third-party-service": "rimraf generated/third-party-service && gen-api-models --api-spec openapi/consumed/api-third-party.yaml --out-dir generated/third-party-service --request-types --response-decoders --client",
     "generate:proxy:remote-content-service": "rimraf generated/remote-content-service && gen-api-models --api-spec openapi/consumed/api_remote_content.yaml --out-dir generated/remote-content-service --request-types --response-decoders --client",
     "generate:proxy:piattaforma-notifiche": "rimraf generated/piattaforma-notifiche && gen-api-models --api-spec openapi/consumed/api-piattaforma-notifiche.yaml --out-dir generated/piattaforma-notifiche --request-types --response-decoders --client",
@@ -87,7 +86,6 @@
     "@pagopa/io-functions-app-sdk": "x",
     "@pagopa/io-functions-cgn-sdk": "x",
     "@pagopa/io-functions-commons": "^29.3.0",
-    "@pagopa/io-functions-eucovidcerts-sdk": "x",
     "@pagopa/ts-commons": "^13.1.2",
     "@pagopa/winston-ts": "^2.2.0",
     "applicationinsights": "^2.9.5",
