+++ src/__tests__/api-specs.test.ts
@@ -17,12 +17,3 @@ describe("API notifications specs", () => {
     expect(api).toBeDefined();
   });
 });
-
-describe("API EUCovidCert specs", () => {
-  const specFilePath = `${__dirname}/../../api_eucovidcert.yaml`;
-
-  it("should be valid", async () => {
-    const api = await SwaggerParser.bundle(specFilePath);
-    expect(api).toBeDefined();
-  });
-});
